refactor operation
